requirements.txt
dnspython==2.6.1
numpy==1.24.4
pandas==2.0.3
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
tzdata==2024.2
